@Carterpersall 2022-10-04
It seems that Windows' reverse geocoding api (System.Device.Location.CivicAddressResolver) does not work within Powershell for whatever reason, which means I can only get the coordinates of the user and not anything else. I tried using the UWP reverse geocoding API (Windows.Services.Maps.MapLocationFinder) too and hit a wall where `FindLocationsAtAsync` was returning an empty `System.__ComObject`, along with it only working in PowerShell 5 for some reason.