@Carterpersall 2023-01-12
> It seems that Windows' reverse geocoding api (`System.Device.Location.CivicAddressResolver`) always fails within Powershell for whatever reason, which means I can only get the coordinates of the user and nothing else. I also tried using the UWP reverse geocoding API (`Windows.Services.Maps.MapLocationFinder`) and hit a wall where `FindLocationsAtAsync` was returning an empty `System.__ComObject` when it should be returning a `MapLocationFinderResult` object, along with it only working in PowerShell 5 for some reason.

Using `System.Device.Location.GeoCoordinateWatcher` works and is quite fast